lib-musl: Add patches for system call handling
In wraperele din musl se executa a binary sistem call, e.g. for x86_64 the `syscall`
instruction is called which does a trap. We don't want this because it is overhead
in a single-address-space OS as Unikraft. Hence, we changed the way musl deals with
syscall handling, by using syscall shim layer's macros which will do a function call instead.

Signed-off-by: Dragos Iulian Argint <dragosargint21@gmail.com>